+++ sdk/storage/azdatalake/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features Added
 * Append API Bundled with Flush functionality
 * HNS Encryption Scope support
+* Append API with acquire lease, release lease and renewal of lease support.
+* Flush API bundled with release lease option.
 
 ### Breaking Changes